requirements.txt
numpy
scipy
pandas
seaborn
natsort
matplotlib

joblib
pyexcel
openpyxl
pyserial
pyfirmata2

shap
xgboost
lightgbm
torchaudio
transformers
scikit-learn
tensorflow
pytorch_wavelets
PyWavelets
tensorly

--no-index
-f https://download.pytorch.org/whl/cu121
torch
torchvision
torchaudio

pyeeg
eeglib
librosa
antropy
torchviz
neurokit2

accelerate
openai-whisper
SpeechRecognition
webdriver_manager

PyQt5
dill
requests
Pillow
ipython
sympy
soundfile
sounddevice
openai
selenium
pygame
baselineremoval
plotapi
